backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email-validator]==2.7.4
httpx==0.27.0
pytesseract>=0.3.10
Pillow==10.4.0
PyMuPDF==1.24.9
openai>=1.35.0
psycopg2-binary==2.9.9
boto3>=1.34.0
cryptography>=42.0.0
